Moved math things to new subpackage, still adapting gochem.
Moved everything to subpackage v3. I am not sure about that name,
especially because I tend to use "v1" and "v2"  as identifiers for
random vectors. Of course it is easy to use search and replace to
change the name of the subpackage.  Fixed goChem files up to
"geometric.go". This file uses unexported functions that are now
in the v3 subpackage so the fixing is more complex.