Added support for v3 in the qm subpackage, Also added v3 dir to repo
I added support to the new scheme in the qm package, including fixing
a bug that had not surfaced before. I also added the v3 dir to the repo,
which I of course should have added a couple of commits ago but forgot
and didn't notice until now, sorry about that.