Added new binaries for PyMOL plugins and minor changes in some pages.
Added new binaries compiled from the devel goChem branch and newAPI
goPyMOL branch. Now only 64 bits binaries are produced and Windows
binaries have been added.

I also made several (minor) modifications to most pages, mostly to
index and imprint (and of course the pymolcomp page). AK Ochsenfeld
does not appear as "goChem is developed here" anymore, but we still
have a link to it as "work place or former workplace of the team".

I added a CC attribution license for the Chemistry Gopher, which is
probably not needed, but still.