API change: Changed the way to specify a QM job.
Before we only had a way for optimize and single-point. There was an Optimize
boolean flag at qm.Calc. Simple but limited and impossible to grow. Now we use
A struct with bool options one of which has to be marked as true. Each QM handler
provides closures to handle each supported job. A nil closure means that the
job is not supported.