Bug fix for clifford.go, it seems to be working but more changes might be needed
There was a bug affecting Rotate, only apparent when a high number of cores was used.
More testing is needed to ensure the program works correctly (including, testing several
different amount of cores). I also commented 2 tests from the suit because of the lack of
the program Reduce in one of the testing machines and because of some missing needed structures
in the testing suite.