The solvation functions have been implemented and tested
The implementation this time is not stupid like the one before
this commit, so it's ~ an order of magnitude faster. I implemented
RDF and MDDF, the solvent can be considered on per-atom basis, or
with the center of mass/centroid.

One implementation for RDF/MDDF goes secuentially through the
trajectory, while the other works in parallel with as many CPUs
as are available in the machine.