BREAKING: The API for the bond path finder changed
Note that this is a newly incorporated function.  Unfortunately,
I did push it to  master a bit too soon, but still.

The new API is cleaner for the most common use case,
and more extensible.

A possible new API change would be to move the path finder
and the ring functions to a subdirectory