Added new option to lovo and fixed a bug.
There was a bug on how the RMSDs were calculated (also, they
were labeled "MSD", even though they were the sq root.
It likely didn't affect the alignment, but the valuest themselves were
wrong. It is fixed now.

Also, you can now say "I want to superimpose all residue with an RMSD
less or equal X", and the function will do its best to get that.
It might involver more iterations, and the limit will be pushed up
if it is not reachable.