Added support for the ctf variant of stf.
ctf is very similar to stf, but it is XYZ-compliant. I.e., a ctf
file is nothing more than an XYZ  file compressed with gzip.
The first frame of the ctf can contain metadata in the "comment" line
of the XYZ (i.e. the data is stored in a XYZ-compatible way).