Modified the python stf to use zstandard
The Python stf implementation now reads
and write zstanrdard-compressed stf files.
It has the issue of requiring a non-standard
library (https://github.com/indygreg/python-zstandard)
But the zstandard compression is faster to read
and write, which is important for us, while
producing files that are only slightly larger
than the deflate-compressed ones. With our settings, of course).
stf files compressed with zstandard use the ".sts" suffix.