Improved the lovo API, and added a function
The LOVO API is now better, I think. The main change
is that the fields of the config struct are no longer
exported. Instead, a set of functions work at the same
time as getter/setters for each field.

I also added a method to format the results in the
format wanted by the Gromacs' gmx make_ndx tool
to build a Gromacs selection with the LOVO-selected
residues.