Voro areas and volumes seem to be working but possible problems.
Given 2 subsets A,B the area considering A->B is not always the same as B->A
This is odd, but it might actually be OK if some of the A/B are unbound
from the point of view of A but not of B, for instance. This might need
fixing, or it might just be part of the errors expected when there are
unbound polyhedra.