Support for Orca 5. Sort-of API break
Several changes, also in the tests, to take better advantage of Orca5.
This involves some changes in the ways calculationss work, including
defaults. This part
has never been considered part of the API, since it always depende on
the QM program, but still, I signal the change as possible break.

One important thing is that low- high- basis are not supported anymore.
I think they make little sense nowadays. The options are still in
qm.Calc, so as not to actually break the API, but it has no effect in
ORCA.