Some gromacs topology file edition functions (untested).
Some functionality to manipulate Gromacs topology files:

Add or remove atoms from  bonded terms (i.e. increase or decrease  the atom number in all
atoms in all bonded terms by a constant). If removing, the bonded terms
involving the removed atoms should have been removed previously

Add terms to a given header (i.e. bonds, angles)

Merge 2 topologies (does not merge includes). This is something to fix,
The reasonable behavior is to merge includes so then the repeated ones
can be removed.

None of this has been tested.

Some of this was actually written a while ago but of course I never
added it to the Git repo, I just realized it now. Those things included
a test but I changed the API so the test doesn't work now.