Started work on a simple clash-resolution system
The idea is to have a system that, given a 'mobile' subsystem and one or
more bonds that can act as axes of rotation for that susbsystem (each
bonds provides a rotation), as well as a topology including vdW radii,
can determine clashes between the mobile system and both itself and
other parts of the system and rotate around the given axes so as to
relieve those clashes.

For this, I have also added some functionality to retrieve vdW radii
from top.FF objects.