Small modification to gc.go and addition to handy.go
in gc.go the function to 'cgize' a backbone now gives the option
to use geometric center (the default behavior still matches the martini3
protein model, using the center of mass for BB beads).

In handy.go there is a new function to rotate a system in such a way
that 2 'axes'/vectors are superimposed. (for instance, a bond with the z
axis).