Fix glitch for topologies, add function
The OneBased attribute of each term in the FF structures
indicates if the indexes are 1 or 0 based. It used to be 0 is 1-based
and 1 if 0-based, which is silly. It's now 1 for 1-based and so on now,
as expected.

I also added a function to add bonds to the *chem.Topology in FF based
on the bonds (and, possibly, constraints) in the itp.