Improvement and bugfix to Gromacs topology r/w
- You have now more freedom to add atoms to your system's FF
only from topology info, not having to have read it from a molecule,
as, in the CG case you might only have the topology available.

- A bugfix: The previous version assumed that improper dihedrals
haver their own section in Gromacs topology. This is NOT the case,
as they are dihedrals with a particular function type. That is fixed
here.