Several changes to bonds and rings
BondedPaths has been replaced with what was BondedPathsFuncs.
so a new item has been added to BondedOptions, a function f(*Bond)bool
that allow skipping bonds for which the function returns false.

A similar thing has been added to FindRings, so now one can exclude
bonds from the search.

An additional MaxRing options has been added to RingOptions, so the max
size allowed for a ring can now be set by the user (it was hardcoded at
6).

Finally, what I think is a bugfix.
FindRings returned, for each ring, the full bonded path. Meaning, the
initial atom appeared at the begining, and also at the end, giving a
path 1 atom longer than the ring. The first element of the path of each
ring is now removed. This is something that could be wrong, so look at
this commit if things start failing.