API BREAK: InWhichRing function changes output.
The InWhichRing function now returns the indexes of all ringst to which
the given atom belongs, or nil. It used to return the index of the first
ring found, or -1.

In addition:
For the Planarity method of rings
you may now give a matrix of the size of the ring to be used as
'scratch' and minimize allocation.