API BREAK: A small detail in CG-izing function
BackBoneCGIze used to take an optional 'centroid' parameter.
It now takes an optional 'com' parameter. Using the center of mass is
still the default, but now, if you want to use the geometric center, you
have to give 'false' as the optional parameter.

Also a function has been added to produce CG coordinates (COM or COG
based) from a molecule, its coordinates and a [][]int where each slice
contains the indexes of the atoms forming the respective bead.