Skip a irrelevant check in the Matrix.SomeVecs method
It used to check whether the set of indexes given (the atoms to take)
are larger than the full matrix, giving a panic if that happens.
It only serves to give a slightly better panic, as you'll still get a
panic if you try to take a row with larger indexes than the rows
present, but it prevents from things like taking the same vector many
times, which you might want to to (a use case appears in Bartender.